Remove columns based on JSON file information
This commit modifies the DBH.py file to add the functionality to delete columns from the database tables.

In the DBIntegrityCheck() function:

    For the SettingsCryptoRates table, the code iterates over the existing columns. If a column is not present in the JSON file and it is not the "chatID" column, it is deleted from the SettingsCryptoRates table.
    Similarly, for the SettingsExchangeRates table, the code iterates over the existing columns. If a column is not present in the JSON file and it is not the "chatID" column, it is deleted from the SettingsExchangeRates table.
    The IgnoredCurrencies table is also checked for columns that do not exist in the JSON file. If any such column is found, it is deleted from the IgnoredCurrencies table.

Additionally, a new function AddIgnoredCurrency(chatID: str) is added to the DBH.py file. This function is used to add the chatID to the IgnoredCurrencies table if it doesn't already exist.

These changes allow for the deletion of columns from the database tables, ensuring the data remains consistent with the JSON file.