Fixed ignored currencies and removed reports
Removed the "Reports" table from the listOfServiceTables as it is no longer needed.
Added functionality to remove tables from the database that are not present in the listOfTables, listOfServiceTables, or listOfStatsTables.
Updated the CreateServiceDataBase function to exclude the creation of the "Reports" table.
Updated the CreateDataBaseTemplate function to set default values for certain columns in the "SettingsExchangeRates" table.
Updated the SetIgnoredCurrency function to accept an integer value for the currency and store it in the "IgnoredCurrencies" table.
Updated the GetIgnoredCurrencies function to return a list of ignored currencies instead of a dictionary.